% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_basic_map.R
\name{basic_map}
\alias{basic_map}
\title{Produce leaflet map to be used as a basis for displaying layers}
\usage{
basic_map()
}
\value{
leaflet map
}
\description{
Produce leaflet map to be used as a basis for displaying layers
}
