% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_regmod.R
\name{calc_regmod}
\alias{calc_regmod}
\title{Produces scatterplot of y against x}
\usage{
calc_regmod(data, x, y, xtrans, ytrans)
}
\arguments{
\item{x}{x variable name in the regional dataset (with quotes)}

\item{y}{y variable name in the regional dataset (with quotes)}
}
\value{
scatterplot
}
\description{
Produces scatterplot of y against x
}
\examples{
calc_regmod(datRMC,
"surface de la bande fluviale",
"surface de la bande active")
}
