% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{define_palette}
\alias{define_palette}
\title{Define palette function for a vector/factor}
\usage{
define_palette(vcolor)
}
\arguments{
\item{vector/factor}{of values to translate into colors}
}
\value{
a palette function
}
\description{
Define palette function for a vector/factor
}
\examples{
pal=define_palette(c("oui","oui","non","peut-être"))
plot(1:3,rep(1,3),pch=20, cex=3,
col=pal(c("oui","non","peut-être")))
pal=define_palette(seq(0,1,0.1))
plot(1:10,rep(1,10),pch=20, cex=3,
col=pal(seq(0,1,0.1)))

}
