% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_available_info}
\alias{get_available_info}
\title{Get available metrics for a river}
\usage{
get_available_info(axis)
}
\arguments{
\item{axis}{river id}
}
\value{
list of metrics families
}
\description{
Get available metrics for a river
}
\examples{
get_available_info("AX0005")
}
