% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_coords}
\alias{get_coords}
\title{Get coords for a river}
\usage{
get_coords(axis)
}
\arguments{
\item{axis}{river id}
}
\value{
coords corresponding to river
}
\description{
Get coords for a river
}
\examples{
get_coords("AX0005")
}
