% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_data}
\alias{get_data}
\title{Get data regarding metric for a river}
\usage{
get_data(axis, zvar)
}
\arguments{
\item{axis}{river id}

\item{zvar}{metric}
}
\value{
tibble with data xvar and zvar
}
\description{
Get data regarding metric for a river
}
\examples{
get_data("AX0005","talweg_slope")
}
