% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_datapath}
\alias{get_datapath}
\title{Get datapath towards file with metric data for a river}
\usage{
get_datapath(axis, zvar)
}
\arguments{
\item{axis}{river id}

\item{zvar}{metric}
}
\value{
path towards metric data for river
}
\description{
Get datapath towards file with metric data for a river
}
\examples{
get_datapath("AX0005","talweg_slope")
}
