% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_info}
\alias{get_info}
\title{Get info (data+labels) regarding metric for a river}
\usage{
get_info(axis, zvar)
}
\arguments{
\item{axis}{river id}

\item{zvar}{metric}
}
\value{
list with data and labels
}
\description{
Get info (data+labels) regarding metric for a river
}
\examples{
get_info(axis="AX0005",zvar="landcover_area")
}
