% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_labels}
\alias{get_labels}
\title{Get labels regarding metric}
\usage{
get_labels(zvar)
}
\arguments{
\item{zvar}{metric}
}
\value{
tibble with labels xlab and ylab
}
\description{
Get labels regarding metric
}
\examples{
get_labels(zvar="landcover_width")
}
