% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_basic_map.R
\name{get_rivers_from_scatterplot}
\alias{get_rivers_from_scatterplot}
\title{Title}
\usage{
get_rivers_from_scatterplot(
  xvar,
  yvar,
  brush_xmin,
  brush_xmax,
  brush_ymin,
  brush_ymax
)
}
\arguments{
\item{xvar}{name of variable x}

\item{yvar}{name of variable y}

\item{brush_xmin}{xmin of considered objects}

\item{brush_xmax}{xmax of considered objects}

\item{brush_ymin}{ymin of considered objects}

\item{brush_ymax}{ymax of considered objects}
}
\value{
river ids
}
\description{
Title
}
