% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_long_data.R
\name{get_shape}
\alias{get_shape}
\title{Get polyline for a river}
\usage{
get_shape(axis)
}
\arguments{
\item{axis}{river id}
}
\value{
polyline corresponding to river
}
\description{
Get polyline for a river
}
\examples{
get_shape("AX0005")
}
