% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_long_profiles.R
\name{mod_long_profiles_ui}
\alias{mod_long_profiles_ui}
\alias{mod_long_profiles_server}
\title{mod_long_profiles_ui and mod_long_profiles_server}
\usage{
mod_long_profiles_ui(id)

mod_long_profiles_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
