% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_regional_models.R
\name{mod_regional_models_ui}
\alias{mod_regional_models_ui}
\alias{mod_regional_models_server}
\title{mod_regional_models_ui and mod_regional_models_server}
\usage{
mod_regional_models_ui(id)

mod_regional_models_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
