% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_historic_profiles}
\alias{plot_historic_profiles}
\title{Plot metric along river}
\usage{
plot_historic_profiles(info, keep_landcover = NA, keep_time = NA)
}
\arguments{
\item{info}{information about axis and z metric: data with columns named lvar and zvar and labels}
}
\value{
a plot showing zvar against lvar
}
\description{
Plot metric along river
}
\examples{
plot_historic_profiles(info=get_info(axis="AX0001", zvar="historic_area"), 
keep_landcover=c("Active Channel","Gravel bars"),
keep_time=c(1938,1990,2019))
}
