% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot metric along river}
\usage{
plot_profiles(dat, y)
}
\arguments{
\item{y}{y-axis label}

\item{data}{data with columns named xvar and yvar}
}
\value{
a plot showing yvar against xvar
}
\description{
Plot metric along river
}
\examples{
plot_profiles(dat=get_data(axis=1, y="LONG_PROFILE"),
              y="height (m)")
}
