% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{plot_regmod}
\alias{plot_regmod}
\title{Produces scatterplot of y against x}
\usage{
plot_regmod(x, y)
}
\arguments{
\item{x}{variable name in the regional dataset (with quotes)}

\item{y}{variable name in the regional dataset (with quotes)}
}
\value{
scatterplot
}
\description{
Produces scatterplot of y against x
}
\examples{
plot_regmod("m_surf_bf","m_lrgm_ba")
}
