% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R
\docType{data}
\name{table_metrics}
\alias{table_metrics}
\title{table documenting possible types of profiles}
\format{A tibble with 4 variables
\describe{
  \item{filename}{corresponding file name}
  \item{varname}{variable name (as stated in the .nc files)}   
  \item{unit}{variable name to be used as y for plotting profile}
  \item{explain}{used as label for x axis when plotting profile}   
  \item{typelzk}{type of variable: l for longitudinal coordinate, z for metric, k for other -e.g. class of landcover-}
  \item{label}{used as label for axes when plotting profiles}
}}
\usage{
table_metrics
}
\description{
Lists possible metrics and corresponding filenames,
as well as corresponding names for x-y variables
in profiles and corresponding labels in plots.
}
\keyword{datasets}
