% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R, R/table_profiles-data.R
\docType{data}
\name{table_profiles}
\alias{table_profiles}
\title{table documenting possible types of profiles}
\format{A tibble with 4 variables
\describe{
  \item{type}{possible metric type (as asked for in the app)}
  \item{filename}{corresponding file name}
  \item{xvar}{variable name to be used as x for plotting profile}   
  \item{yvar}{variable name to be used as y for plotting profile}
}}
\usage{
table_profiles

table_profiles

table_profiles
}
\description{
Lists possible metrics and corresponding filenames,
as well as corresponding names for x-y variables
in profiles.

Lists possible metrics and corresponding filenames,
as well as corresponding names for x-y variables
in profiles.

Lists possible metrics and corresponding filenames,
as well as corresponding names for x-y variables
in profiles.
}
\keyword{datasets}
