% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-data.R
\docType{data}
\name{table_regmod}
\alias{table_regmod}
\title{table documenting possible types of regional models}
\format{A tibble with variables
\describe{
  \item{name}{possible metric type (as asked for in the app)}
  \item{label}{corresponding file name}
}}
\usage{
table_regmod
}
\description{
Lists possible variable names and corresponding labels.
}
\keyword{datasets}
