% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_basic_map.R
\name{add_rivers_to_map}
\alias{add_rivers_to_map}
\title{Title}
\usage{
add_rivers_to_map(
  map,
  sfdata,
  group = "base",
  vlayerId = "ID",
  vcolor = NA,
  color = "grey",
  vpopup = "popup"
)
}
\arguments{
\item{map}{a leaflet map}

\item{sp}{the polylines spatial data corresponding to rivers that will be added to the leaflet map}
}
\value{
map with river lines
}
\description{
Title
}
\examples{
basic_map() \%>\%
 add_rivers_to_map(sfdata=datsf ,vcolor=NA,vpopup="ID")
}
