% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_basic_map.R
\name{get_rect_bounds_from_profile}
\alias{get_rect_bounds_from_profile}
\title{Get bounds of rectangle (on map) based on rectangle on plot}
\usage{
get_rect_bounds_from_profile(obj_id, lmin, lmax)
}
\arguments{
\item{obj_id}{river id}

\item{lmin}{minimum distance along axis}

\item{lmax}{maximum distance along axis}
}
\value{
data with min-max of lng and min-max of lat
}
\description{
Get bounds of rectangle (on map) based on rectangle on plot
}
