% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_regmod.R
\name{plot_regmod}
\alias{plot_regmod}
\title{Produces scatterplot of y against x}
\usage{
plot_regmod(data_mod)
}
\arguments{
\item{data_mod}{a model}
}
\value{
scatterplot
}
\description{
Produces scatterplot of y against x
}
\examples{
plot_regmod(calc_regmod(datRMC \%>\% sf::st_drop_geometry(),
 "surface de la bande fluviale",
 "surface de la bande active"))
}
